Revert "Make maven build independent of http://sun.java.com. Remove internet availability dependency"
The commit 4ece39 allows jetty:run to be able to run the ICE instance
without the need for an internet connection. However, it breaks tomcat
deployment, as runtime location of the file cannot be specified ahead
of time. And since the use cases for running ICE without an internet
connection seems pretty minimal, this "feature" is being reversed.

This reverts commit 4ece3966d96829b227e8d3607d0d504bd8e1f000.