Add error handling to email error logs
ICE was not doing any error handling around the logging facility
sending emails on errors. If there were any errors in sending email
on an an error, the exception would cause the original error to be
lost. This change wraps the error handling code in try-catch blocks
to add additional logging when emails cannot be sent, and letting
the original error reach the log files.