Prevent stack overflow in error email handler (#73)
The Logger class will send emails to the instance admin for all calls to
Logger.error(). If there is an error in sending this email, there is an
infinite recursion loop ending with a stack overflow. Modify the Logger
class to add a Logger.errorLocalOnly() call that will *not* send an email.
The CustomEmail class will use this local-only handler when encountering
an error sending mail, eliminating the infinite recursion.