Update POM build plugin versions
When I build on a fresh maven install, executing `mvn install` with the
previous POM would fail with errors in the `default-war` goal in the `package`
stage. There would be a warning that looks like this:

```
[WARNING] Error injecting: org.apache.maven.plugin.war.WarMojo
com.google.inject.ProvisionException: Unable to provision, see the following errors:

1) Error injecting constructor, java.lang.ExceptionInInitializerError: Cannot access defaults field of Properties
  at org.apache.maven.plugin.war.WarMojo.<init>(Unknown Source)
  while locating org.apache.maven.plugin.war.WarMojo

*** redacted several lines of stack trace ***
```

And just before the end of the build, this:

```
[ERROR] Failed to execute goal org.apache.maven.plugins:maven-war-plugin:2.2:war (default-war) on project ice: Execution default-war of goal org.apache.maven.plugins:maven-war-plugin:2.2:war failed: Unable to load the mojo 'war' in the plugin 'org.apache.maven.plugins:maven-war-plugin:2.2' due to an API incompatibility: org.codehaus.plexus.component.repository.exception.ComponentLookupException: Cannot access defaults field of Properties
```

To keep compatibility with latest Maven and OpenJDK, explicitly adding the WAR
plugin latest version fixes the issue. The error message indicates that version
2.2 was being used, while the current latest is version 3.3.1.

Also moving the explicit version of the Maven Compiler Plugin from the POM
dependencies section, to the build plugins section. Putting the WAR Plugin in
dependencies has no effect, while putting it in build plugins works; so it
stands to reason that explicitly versioning the Compiler Plugin in dependencies
has no effect, and its proper place is also in build plugins.