Update Docker base image version
There is no longer an OpenJDK 13 based image for Maven. Switching instead to a
newer OpenJDK 16 based image. The `3-openjdk-16` image is also tagged with
`latest` on Docker Hub.

Changing the base Tomcat image from tag `9` to `9-jdk16`. Using a WAR compiled
with an OpenJDK 16 image will not launch correctly on the OpenJDK 11 JRE that's
used in the plain `9` tag image.