Refresh docker (#111)
* Update POM build plugin versions

When I build on a fresh maven install, executing `mvn install` with the
previous POM would fail with errors in the `default-war` goal in the `package`
stage. There would be a warning that looks like this:

```
[WARNING] Error injecting: org.apache.maven.plugin.war.WarMojo
com.google.inject.ProvisionException: Unable to provision, see the following errors:

1) Error injecting constructor, java.lang.ExceptionInInitializerError: Cannot access defaults field of Properties
  at org.apache.maven.plugin.war.WarMojo.<init>(Unknown Source)
  while locating org.apache.maven.plugin.war.WarMojo

*** redacted several lines of stack trace ***
```

And just before the end of the build, this:

```
[ERROR] Failed to execute goal org.apache.maven.plugins:maven-war-plugin:2.2:war (default-war) on project ice: Execution default-war of goal org.apache.maven.plugins:maven-war-plugin:2.2:war failed: Unable to load the mojo 'war' in the plugin 'org.apache.maven.plugins:maven-war-plugin:2.2' due to an API incompatibility: org.codehaus.plexus.component.repository.exception.ComponentLookupException: Cannot access defaults field of Properties
```

To keep compatibility with latest Maven and OpenJDK, explicitly adding the WAR
plugin latest version fixes the issue. The error message indicates that version
2.2 was being used, while the current latest is version 3.3.1.

Also moving the explicit version of the Maven Compiler Plugin from the POM
dependencies section, to the build plugins section. Putting the WAR Plugin in
dependencies has no effect, while putting it in build plugins works; so it
stands to reason that explicitly versioning the Compiler Plugin in dependencies
has no effect, and its proper place is also in build plugins.

* Update Docker base image version

There is no longer an OpenJDK 13 based image for Maven. Switching instead to a
newer OpenJDK 16 based image. The `3-openjdk-16` image is also tagged with
`latest` on Docker Hub.

Changing the base Tomcat image from tag `9` to `9-jdk16`. Using a WAR compiled
with an OpenJDK 16 image will not launch correctly on the OpenJDK 11 JRE that's
used in the plain `9` tag image.

* Add cache config to Docker Tomcat config.xml

Without this, during Tomcat/ICE startup in a container, the logs would be
flooded with messages like this (times about 400):

```
WARNING [main] org.apache.catalina.webresources.Cache.getResource Unable to add the resource at [/WEB-INF/classes/ice-5.9.4/WEB-INF/lib/commons-codec-1.9.jar] to the cache for web application [] because there was insufficient free space available after evicting expired cache entries - consider increasing the maximum size of the cache
```

Adding cache config that's about 10x the default size makes the problem go
away, and vastly improves startup times -- a little over 15 seconds to startup
without the config change, vs a little under 5 seconds with the change.

* Move default database parameters to CATALINA_OPTS

Previous methods of injecting database credentials were not working correctly
anymore, so adding the ability to configure the values, while keeping defaults.